+++ plugins/__init__.py
@@ -19,6 +19,7 @@ from .steam_lang_codes import (
 )
 from .auto_updater import AutoUpdater
 from .icon_loader import IconLoader
+from .http_client import HTTPClient
 
 __all__ = [
     'HighlightDelegate',
@@ -39,5 +40,6 @@ __all__ = [
     'get_display_name',
     'get_code_from_display_name',
     'AutoUpdater',
-    'IconLoader'
+    'IconLoader',
+    'HTTPClient'
 ]